bueno hubo avance nms falta algunas restricciones y los detalle de pago efectivo y de recarga saldo... y los botones de atras

+++ TPAGO_APP2/app/src/main/java/com/example/tpago2/adapter/TarjetaAdapter.kt
@@ -4,6 +4,8 @@ import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.Button
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
 import com.example.tpago2.R
 import com.example.tpago2.data.dao.TarjetaUsuarioDAO
@@ -33,13 +35,27 @@ class TarjetaAdapter(private val items: MutableList<TarjetaUsuario>, private val
     }
 
     private fun deleteTarjeta(index: Int, item: TarjetaUsuario) {
-        val daoTarjeta = TarjetaUsuarioDAO(context2)
-        daoTarjeta.eliminarTarjetaUsuario(item.num_tarjeta)
+        val builder = AlertDialog.Builder(context2)
+        builder.setTitle("Confirmación")
+        builder.setMessage("¿Estás seguro de que deseas eliminar esta tarjeta?")
+        builder.setPositiveButton("Sí") { _, _ ->
+            // Acción a realizar si el usuario confirma la eliminación
+            val daoTarjeta = TarjetaUsuarioDAO(context2)
+            daoTarjeta.eliminarTarjetaUsuario(item.num_tarjeta)
 
-        items.removeAt(index)
-        notifyItemRemoved(index)
-        notifyItemRangeChanged(index, items.size)
+            items.removeAt(index)
+            notifyItemRemoved(index)
+            notifyItemRangeChanged(index, items.size)
+            Toast.makeText(context2, "Tarjeta eliminada correctamente", Toast.LENGTH_SHORT).show()
+        }
+        builder.setNegativeButton("Cancelar") { dialog, _ ->
+            // Acción a realizar si el usuario cancela la eliminación
+            dialog.dismiss()
+        }
+        val dialog = builder.create()
+        dialog.show()
     }
+
 
     fun editTarjeta(index: Int, tarjeta: TarjetaUsuario) {
         items[index] = tarjeta
